+++ R/expand.R
@@ -26,7 +26,7 @@
 #'   For factors, the full set of levels (not just those that appear in the
 #'   data) are used. For continuous variables, you may need to fill in values
 #'   that don't appear in the data: to do so use expressions like
-#'   `year = 2010:2020` or `year = \link{full_seq}(year)`.
+#'   `year = 2010:2020` or `year = \link{full_seq}(year,1)`.
 #'
 #'   Length-zero (empty) elements are automatically dropped.
 #' @seealso [complete()] for a common application of `expand`:
@@ -150,8 +150,8 @@ cross_df <- function(x, y) {
   y_idx <- rep(seq_len(nrow(y)), nrow(x))
   dplyr::bind_cols(x[x_idx, , drop = FALSE], y[y_idx, , drop = FALSE])
 }
-drop_empty <- function(x) {
-  empty <- map_lgl(x, function(x) length(x) == 0)
+drop_empty <- function(x, factor = TRUE) {
+  empty <- map_lgl(x, function(x) length(x) == 0 & (!factor | !is.factor(x)))
   x[!empty]
 }
 
@@ -161,7 +161,7 @@ nesting <- function(...) {
   x <- tibble::lst(...)
 
   stopifnot(is_list(x))
-  x <- drop_empty(x)
+  x <- drop_empty(x, factor = FALSE)
 
   df <- as_tibble(x)
   df <- dplyr::distinct(df)